o Make ocfs2_extend_allocation() try to clean up upon error. o Put a debug chunk in extend_file.c to extend inodes.
Original-revision: 478

+++ libocfs2/alloc.c
@@ -389,7 +389,7 @@ out:
 }
 
 errcode_t ocfs2_free_clusters(ocfs2_filesys *fs,
-			      uint64_t len,
+			      uint32_t len,
 			      uint64_t start_blkno)
 {
 	errcode_t ret;
